Promote NcsCodedAttribute to top level. #3633.
It will shortly have much more to do.
