+++ lib/ncs_navigator/core/mdes.rb
@@ -6,9 +6,10 @@ module NcsNavigator::Core
     autoload :CodeListLoader, 'ncs_navigator/core/mdes/code_list_loader'
 
     autoload :HumanReadablePublicIdGenerator, 'ncs_navigator/core/mdes/human_readable_public_id_generator'
-    autoload :InstrumentOwner, 'ncs_navigator/core/mdes/instrument_owner'
-    autoload :MdesDate,        'ncs_navigator/core/mdes/mdes_date'
-    autoload :MdesRecord,      'ncs_navigator/core/mdes/mdes_record'
+    autoload :InstrumentOwner,   'ncs_navigator/core/mdes/instrument_owner'
+    autoload :MdesDate,          'ncs_navigator/core/mdes/mdes_date'
+    autoload :MdesRecord,        'ncs_navigator/core/mdes/mdes_record'
+    autoload :NcsCodedAttribute, 'ncs_navigator/core/mdes/ncs_coded_attribute'
 
     autoload :Version,           'ncs_navigator/core/mdes/version'
     autoload :VersionMigrator,   'ncs_navigator/core/mdes/version_migrator'
